fix(volcminer): desloppify
